frametrim: split OpenGL implementation into a new class
To allow implementations for different graphics API, the OpenGL
has been moved to a new class (frametrimmer::OpenGLImpl). The
FrameTrimmer class now only contains the most generic methods
and callback caching.
